requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
openai
numpy
sentence_transformers
python-dotenv